+++ app/(nav)/placeRegister/page.tsx
@@ -2,56 +2,19 @@
 
 import TopBar from '@/components/Common/TopBar/TopBar';
 import PlaceRegister from '@/components/Layout/PlaceRegister/PlaceRegister';
+import Button from '@/components/Common/Button/Button';
+
 import {
   ButtonLayout,
   PlaceLayout,
   PlaceRegisterWrapper,
 } from './PlaceRegister.styles';
 import { PageWrapper } from '@/app/commonPage.styles';
-import { useEffect, useState } from 'react';
-import Button from '@/components/Common/Button/Button';
+
+import useUploadFiles from '@/hooks/placeRegister/useUploadFiles';
 
 const PlaceRegisterPage = () => {
-  const [postFiles, setPostFiles] = useState<File[]>([]);
-  const [postInfo, setPostInfo] = useState<object>({
-    title: '',
-    hint: '',
-  });
-
-  useEffect(() => {
-    console.log(postInfo);
-  }, [postInfo]);
-
-  const uploadFile = () => {
-    console.log(postFiles);
-
-    const formData = new FormData();
-
-    // 이미지 등록
-    postFiles.map((file) => {
-      formData.append('files', file);
-    });
-
-    // 장소명, 힌트 등록
-    const locationInfoBlob = new Blob([JSON.stringify(postInfo)], {
-      type: 'application/json',
-    });
-    formData.append('locationInfo', locationInfoBlob);
-
-    console.log(Array.from(formData));
-
-    //axios post
-    // axios.post('http://localhost:3079/file/uploads', formData, {
-    //   headers: {
-    //     'Content-Type': 'multipart/form-data'
-    //   }
-    // })
-    // .then((res) => {
-    //     console.log(res.data);
-    // }).catch((err) => {
-    //     console.error(err);
-    // });
-  };
+  const { setPostFiles, setPostInfo, uploadFile } = useUploadFiles();
 
   return (
     <>
